pyproject.toml
[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[project]
name = "style-bert-vits2"
version = "3.0.0.dev0"
description = "Style-Bert-VITS2: Bert-VITS2 with more controllable voice styles."
readme = "README.md"
requires-python = ">=3.10"
license = "AGPL-3.0-only"
keywords = []
authors = [
  { name = "litagin02", email = "139731664+litagin02@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "e2k @ git+https://github.com/tsukumijima/e2k.git@master",
  "jaconv",
  "loguru",
  "num2words",
  "numba",
  "numpy",
  "onnxruntime",
  "pydantic",
  "pyopenjtalk-plus",
  "pyworld-prebuilt",
  "safetensors",
  "transformers",
]

[project.urls]
Documentation = "https://github.com/litagin02/Style-Bert-VITS2#readme"
Issues = "https://github.com/litagin02/Style-Bert-VITS2/issues"
Source = "https://github.com/litagin02/Style-Bert-VITS2"

[project.optional-dependencies]
cpu = ["torch>=2.9", "torchaudio>=2.9"]
cu128 = ["torch>=2.9", "torchaudio>=2.9"]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu128" }]]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchaudio = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "gradio<6.0",
    "ipykernel",
    "librosa",
    "matplotlib",
    "onnx",
    "onnxconverter-common",
    "onnxscript",
    "onnxsim-prebuilt",
    "pyannote-audio==4.0.1",  # 4.0.2以降は torch==2.8.0 と pin されており 2.9 以降と競合するため
    "pyloudnorm",
    "pytest",
    "rich",
    "scipy",
    "soundfile",
    "tensorboard",
    "tqdm",
    "umap-learn",
]


[tool.uv.build-backend]
# パッケージはリポジトリ直下にあるので src レイアウトではなく root を使う
module-name = "style_bert_vits2"
module-root = ""
# もともとの hatch sdist only-include を uv の include に寄せたもの
source-include = [
  ".vscode/**",
  "dict_data/**",
  "docs/**",
  "tests/**",
  "LGPL_LICENSE",
  "LICENSE",
  "README.md",
]
source-exclude = [".git", ".gitignore", ".gitattributes"]

[tool.coverage.run]
source_pkgs = ["style_bert_vits2", "tests"]
branch = true
parallel = true
omit = ["style_bert_vits2/constants.py"]

[tool.coverage.paths]
style_bert_vits2 = ["style_bert_vits2", "*/style-bert-vits2/style_bert_vits2"]
tests = ["tests", "*/style-bert-vits2/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
lines-after-imports = 2
